requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
requests>=2.31.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0
pydantic>=2.4.0
uuid6>=2023.5.2
PyYAML>=6.0.1
